requirements.txt
asyncio
aiofiles
loguru
python-dotenv
retry
openpyxl
urllib3
PyExecJS
requests
argparse
websockets
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
